Django/fileload/requirements.txt
APScheduler==3.9.1	
Django==4.1.1	
PyMySQL==1.0.2	
SQLAlchemy==1.4.41	
arrow==1.2.3	
asgiref==3.5.2	
certifi==2022.9.24	
charset-normalizer==2.1.1	
et-xmlfile==1.1.0	
greenlet==1.1.3	
idna==3.4	
mysqlclient==2.1.1	
numpy==1.23.3	
openpyxl==3.0.10	
pandas==1.5.0	
pip==21.2.3	
python-dateutil==2.8.2	
pytz==2022.2.1	
pytz-deprecation-shim==0.1.0.post0	
requests==2.28.1	
setuptools==57.4.0	
six==1.16.0	
sqlparse==0.4.2	
tzdata==2022.2	
tzlocal==4.2	
urllib3==1.26.12	
xlrd==2.0.1	
djangorestframework==3.14.0